refactor: Simplify token handling and improve test structure in setup
- Refactored token generation and extraction logic into separate functions for better modularity and readability.
- Updated test setup to utilize new helper functions for creating mock requests and responses, enhancing maintainability.
- Removed redundant mock implementations in favor of centralized helper functions, streamlining the test code.
